refactor(design-system): align Apod design elements with MaterialX
This commit updates the Apod design system components, including
IconsData, RadiusData, SpacingsData, BoxShadowsData, and DurationsData,
to align with the MaterialX library standards. This migration enhances
consistency with modern UI design principles, ensures compatibility
with MaterialX guidelines, and facilitates a more cohesive and scalable
design language across the application.

+++ design_system/lib/src/base/durations/x_standard_milliseconds.dart
@@ -1,17 +1,17 @@
 abstract final class XStandardMilliseconds {
-  static const double x100 = 100;
-  static const double x200 = 200;
-  static const double x300 = 300;
-  static const double x500 = 500;
-  static const double x800 = 800;
-  static const double x1000 = 1000;
-  static const double x1200 = 1200;
-  static const double x1500 = 1500;
-  static const double x2000 = 2000;
-  static const double x3000 = 3000;
-  static const double x5000 = 5000;
-  static const double x10000 = 10000;
-  static const double x20000 = 20000;
-  static const double x30000 = 30000;
-  static const double x60000 = 60000;
+  static const int x100 = 100;
+  static const int x200 = 200;
+  static const int x300 = 300;
+  static const int x500 = 500;
+  static const int x800 = 800;
+  static const int x1000 = 1000;
+  static const int x1200 = 1200;
+  static const int x1500 = 1500;
+  static const int x2000 = 2000;
+  static const int x3000 = 3000;
+  static const int x5000 = 5000;
+  static const int x10000 = 10000;
+  static const int x20000 = 20000;
+  static const int x30000 = 30000;
+  static const int x60000 = 60000;
 }
